requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
Django==4.0.1
django-bootstrap4==21.2
django-filter==21.1
django-tables2==2.4.1
djangorestframework==3.13.1
importlib-metadata==4.10.1
Markdown==3.3.6
numpy==1.22.1
pandas==1.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
zipp==3.7.0
